insta_cloan_task/requirements.txt
alembic==1.10.3
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ecdsa==0.18.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.6.2
Flask==2.2.3
Flask-API==3.0.post1
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.2.0
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
PyJWT==2.6.0
pytest==7.3.1
pytest-bdd==6.1.1
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.9
tomli==2.0.1
typing-extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
